pyproject.toml
[project]
name = "mvp"
version = "0.1.0"
description = ""
authors = [
    {name = "Mateus Kasuya",email = "mateusvbkasuya@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo (>=4.11,<5.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
